+++ releasenotes.md
@@ -26,7 +26,8 @@
 - Align literal types with C semantics.
 - Use value promotion instead of signedness promotion to int. So that small unsigned types promote to int, not uint.
 - Add a `@mustinit` attribute to enforce zero-initialization of a type. #3094
-
+- Improve error message when keyword is used instead of an expression. #3088
+ 
 ### Stdlib changes
 - `std::collections::RingBuffer` is renamed `RingList`.
 - Add `List.remove_unordered_at`.